requirements.txt
pandas==1.5.2
Django
django-cors-headers
djangorestframework
